Consolidate ThemeColor definitions
We used to define the theme color twice, once in hex and once in hsl.
To prevent duplication, I've changed the theme color definition to only
use hsl.

Because HSL is not very intuitive, I've also added a link to a website
where you can can create a hsl theme color.